pyproject.toml
[tool.pdm]

[project]
name = "for-nursery"
version = "0.1"
description = ""
authors = []
dependencies = [
    "Cython>=0.29.34",
    "fastapi>=0.100.0",
    "uvicorn[standard]>=0.22.0",
    "strawberry-graphql[fastapi]>=0.194.4",
    "gunicorn>=20.1.0",
    "psycopg[binary]>=3.1.9",
    "SQLAlchemy>=2.0.18",
    "pydantic>=2.0.2",
    "pydantic-settings>=2.0.2",
]
requires-python = ">=3.11"
license = {text = "MIT"}

[tool.pdm.dev-dependencies]
lint = [
    "black>=23.3.0",
    "isort>=5.12.0",
    "flake8>=6.0.0",
    "bandit[toml]>=1.7.5",
    "pyproject-flake8>=6.0.0.post1",
    "mdformat-gfm>=0.3.5",
    "mypy>=1.4.1",
]

test = [
    "Faker>=18.13.0",
    "pytest>=7.4.0",
    "pytest-postgresql>=5.0.0",
    "devtools[pygments]>=0.11.0",
]

[tool.mypy]
plugins = [
  "pydantic.mypy",
  "sqlalchemy.ext.mypy.plugin"
]

follow_imports = "silent"
warn_redundant_casts = true
warn_unused_ignores = true
disallow_any_generics = true
check_untyped_defs = true
no_implicit_reexport = true
# ignore_missing_imports = true
# for strict mypy: (this is the tricky one :-))
disallow_untyped_defs = true
exclude = "/tests/"

[tool.pydantic-mypy]
init_forbid_extra = true
init_typed = true
warn_required_dynamic_aliases = true

[tool.black]
line-length = 88

[tool.isort]
profile = "black"

[tool.flake8]
# black に準拠
max-line-length = 88

[tool.pdm.scripts]
dev = "uvicorn src.app.main:app --reload --port 8080"
server = "gunicorn src.app.main:app -w 4 -k uvicorn.workers.UvicornWorker"
lint = "pflake8 src"
isort = "isort ."
black = "black src"
mdformat = "mdformat ."
format = {composite = ["black", "isort", "mdformat"]}
seed = "src/app/common/utils/seed.py"
